requirements.txt
faiss-cpu
numpy
pymongo
